Sidebar rail-column geometry: pin icons to the slim centerline in both states
Icon controls and the level-0 content column now derive from
--theme-layout-sidebar-slim-width / 2 in the expanded state too, so nothing
moves horizontally while the width animates (measured: the previous instant
flex-center swap teleported icons ~110px to the still-wide panel's midline
and glid them back over 300ms). Slim mode reuses the identical left-anchored
geometry, so centering stays token-exact without justify-center.

Trade-offs: the expanded icon/text column moves from inset+12 to slim/2-8px
(20 -> 24 at defaults) and is now coupled to the slim-width token; mixed-size
icon controls (toggle, avatar) align by center rather than left edge.

New I5 e2e invariant records per-frame icon trajectories during toggle and
asserts max 1px frame movement and equal resting positions across states;
measurement helpers moved to e2e-tests/tests/sidebarGeometry.ts.

Isolated commit on purpose: revert this alone to fall back to flex-center
slim centering.
